EeriecastDjango/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-extensions==3.2.3
django-unfold==0.59.0

# Database
psycopg2-binary==2.9.9

# File handling and storage
pillow==10.0.1
whitenoise==6.6.0
boto3==1.34.162

# Background tasks
celery==5.3.4
redis==5.0.1

# AI/ML integrations
openai==1.3.5

# Payments
stripe==8.10.0

# Utilities
python-decouple==3.8
requests==2.31.0
pytz==2025.2
feedparser==6.0.11

# Production server
gunicorn==21.2.0

# Core Python dependencies
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
idna==3.10
jwcrypto==1.5.6
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3>=1.25.4,<2.1

# OAuth and API extensions
django-oauth-toolkit==3.0.1
drf-nested-routers==0.93.5
